fix(can): fix negative temperature encoding/decoding in ECU→Clocks pipeline
ECU was casting temperature floats directly to uint8_t, which is
undefined behaviour for negative values and resulted in garbled data
(e.g. -10°C → 0 on RP2040). The Clocks side also failed to sign-extend
the fuel temperature byte on receive.

Changes:
- ECU/can.c: replace bare (uint8_t) casts in CAN_updaterecipients_01()
  for coolant, oil, intake and fuel temperature fields by using hal_can_encode_temp_i8() 

refactor(timers): migrate ECU, Clocks, OilAndSpeed to HAL timer table API

Replace manual timer arrays and for-loops with hal_soft_timer_setup_table()
/ hal_soft_timer_tick_table() added in JaszczurHAL in a prior commit.

- ECU/start.c: use hal_soft_timer_table_entry_t + setup/tick table
- Clocks/logic.cpp: use hal_soft_timer_table_entry_t + setup/tick table
- OilAndSpeed/start.cpp: use hal_soft_timer_table_entry_t + setup/tick table

fix(sensors): fix SPI CS pin contention causing CAN init failure on startup

- ECU/sensors.c: configure SPI CS pin correctly to prevent bus conflict
  between SPI peripherals at startup

+++ src/Clocks/can.cpp
@@ -60,8 +60,8 @@ static void onCanFrame(uint32_t canID, uint8_t len, const uint8_t *buf) {
       valueFields[F_CALCULATED_ENGINE_LOAD] = buf[CAN_FRAME_ECU_UPDATE_ENGINE_LOAD];
       valueFields[F_VOLTS] = decToFloat(buf[CAN_FRAME_ECU_UPDATE_VOLTS_HI],
                                           buf[CAN_FRAME_ECU_UPDATE_VOLTS_LO]);
-      valueFields[F_COOLANT_TEMP] = buf[CAN_FRAME_ECU_UPDATE_COOLANT];
-      valueFields[F_OIL_TEMP] = buf[CAN_FRAME_ECU_UPDATE_OIL];
+      valueFields[F_COOLANT_TEMP] = static_cast<int8_t>(buf[CAN_FRAME_ECU_UPDATE_COOLANT]);
+      valueFields[F_OIL_TEMP] = static_cast<int8_t>(buf[CAN_FRAME_ECU_UPDATE_OIL]);
     }
     break;
 
@@ -103,7 +103,7 @@ static void onCanFrame(uint32_t canID, uint8_t len, const uint8_t *buf) {
     case CAN_ID_ECU_UPDATE_02: {
       ecuMessages++; ecuConnected = true;
 
-      valueFields[F_INTAKE_TEMP] = buf[CAN_FRAME_ECU_UPDATE_INTAKE];
+      valueFields[F_INTAKE_TEMP] = static_cast<int8_t>(buf[CAN_FRAME_ECU_UPDATE_INTAKE]);
       valueFields[F_FUEL] = MsbLsbToInt(buf[CAN_FRAME_ECU_UPDATE_FUEL_HI],
                                         buf[CAN_FRAME_ECU_UPDATE_FUEL_LO]);
       valueFields[F_GPS_IS_AVAILABLE] = buf[CAN_FRAME_ECU_UPDATE_GPS_AVAILABLE];
@@ -115,7 +115,7 @@ static void onCanFrame(uint32_t canID, uint8_t len, const uint8_t *buf) {
       ecuMessages++; ecuConnected = true;
 
       valueFields[F_PRESSURE_PERCENTAGE] = buf[CAN_FRAME_ECU_UPDATE_PRESSURE_PERCENTAGE];
-      valueFields[F_FUEL_TEMP] = buf[CAN_FRAME_ECU_UPDATE_FUEL_TEMP];
+      valueFields[F_FUEL_TEMP] = static_cast<int8_t>(buf[CAN_FRAME_ECU_UPDATE_FUEL_TEMP]);
       valueFields[F_FAN_ENABLED] = buf[CAN_FRAME_ECU_UPDATE_FAN_ENABLED];
     }
     break;
